media: imx: imx8-isi: Fix format enumeration
The format enumeration ioctl is only supposed to offer formats which are
supported by the whole system configuration and not just return all
possible formats the capture interface supports.

Thus get the current format from the sensor subdevice and only return
this format in case it matches a supported format of the capture
interface.

Signed-off-by: Stefan Riedmueller <s.riedmueller@phytec.de>
Signed-off-by: Christian Hemp <c.hemp@phytec.de>