staging: media: imx: isi: Remove dependency from ISI core to cap & m2m
There's a circular dependency between the ISI core and the cap and m2m
drivers. The latter use the API implemented by the core, and the core
calls back into the cap and m2m drivers to handle frame end interrupts.

Break this dependency by implementing frame end signalling through a
function pointer. The implementation is probably racy at this point, but
it's no worse than the existing code.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>