media: i2c: imx290: Support variable-sized registers
The IMX290 has registers of different sizes. To simplify the code,
handle this in the read/write functions instead of in the callers by
encoding the register size in the symbolic name macros. All registers
are defined as 8-bit for now, a subsequent change will move to larger
registers where applicable.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
---
Changes since v1:

- Use IMX290_REG_SIZE_SHIFT in IMX290_REG_*BIT() macros
- Use unsigned integers in IMX290_REG_*BIT() macros