arm64: dts: verdin-imx8mp: disable broken energy efficient ethernet eee
Real Energy Efficient Ethernet with Tx LPI (only supported on the
on-module EQOS based MAC) may cause issues with EEE enabled
peers/switches and is therefore disabled in NXP's BSPs (only for Gigabit
even though this is also an issue for FastEthernet!). While NXP may have
patches to fix this we also just disable it for now.

Related-to: ELB-3757
Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>