staging: media: imx: isi: Fix media bus code handling in .set_fmt()
Restrict the media bus code on the sink pad of the ISI subdev to
supported formats, and propagate it to the source pad. The source pad
code is currently restricted to the same code as on the sink pad. This
constraint should be relaxed when adding support for the color space
conversion of the ISI processing pipeline.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>