staging: media: imx: isi: Right-align 10- and 12-bit formats
The ISI shifts the 10-bit and 12-bit formats left by 6 and 4 bits when
using MXC_ISI_OUT_FMT_RAW10 or MXC_ISI_OUT_FMT_RAW12 respectively, to
align the bits to the left and pad with zeros in the LSBs. The
corresponding V4L2 formats are however right-aligned, we have to use
MXC_ISI_OUT_FMT_RAW16 to avoid the left shift.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>