media: mc: entity: Require .has_route() to take a sink and source
The media entity .has_route() operation accepts any two pads of an
entity. It is ill-defined if the pads are both sinks or sources. While
we could clarify the behaviour of the function in those cases, it would
be difficult to implement for drivers, and there are no existing or
foreseen use cases for this. Require the operation to be given a sink
and a source pad instead.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>