dt-bindings: media: imx8-isi: Simplify clocks
There is no need to specify both the root and non-root display AXI and
APB clocks, as the former are children of the latter, and the non-root
clocks are not routed to the ISI. Simplify the clocks by dropping the
non-root clocks, and removing the disp_ prefix and _root suffix.

There may be a need to specify the per-channel clocks if some devices
have the ability to control them independently. Add a corresponding TODO
comment.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>