staging: media: imx: isi: Don't store subdev format info
Get the format info at runtime instead of storing it in the subdev. This
allows dropping the mxc_isi_frame structure at the expense of looking up
format info when starting streaming. If this becomes a problem, the V4L2
subdev state infrastructure should be extended to allow extending the
state with driver-specific data.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>