media: rkisp1: Add helper to check if an internal CSI RX is available
The ISP in the i.MX8MP SoC has the same IP core as the rockchip isp1.
Some versions of this ISP have a built-in CSI receiver and don't need an
external one, such as the one in the rk3399, whereas the one in the
i.MX8MP needs and external one. Add a simple helper based on ISP version
that determines whether or not an internal CSI RX is available.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>