media: rkisp1: capture: Fix and simplify (un)registration
The rkisp1_register_capture() and rkisp1_unregister_capture() functions
don't destroy the mutex (in the error path for the former). Fix this and
make rkisp1_unregister_capture() and rkisp1_capture_devs_unregister()
safe to be called on an unregistered capture node to prepare for
simplification of error handling at probe time.

While at it, drop the double initialization of cap->rkisp1 in
rkisp1_capture_devs_register() as the field is already initialized in
rkisp1_capture_init().

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
---
Changes since v2:

- Use video_is_registered() as registration test
- Drop double initialization of cap->rkisp1
- Move part of 03/17 to this patch

Changes since v1:

- Reset cap->rkisp1 in rkisp1_capture_devs_register()