media: imx: isi: Disable chained pipe in free()
The chained pipe should be reset at resource free time.

This is particularly relevant for the m2m module, as
mxc_isi_channel_free() needs to be called once the last chained
context is released, which might or might not coincide
with the time the last overall user terminates.

For this reason, releasing the chained channel at channel_free() time
allows to use the secondary channel if the only remaining m2m user does not
require chaining.

Signed-off-by: Jacopo Mondi <jacopo@jmondi.org>