[HACK] media: imx: imx8-isi: Add BSP-specific clocks and resets
The media-blk-ctrl handles resets and clocks differently between the
mainline and the BSP implementations of the corresponding drivers. In
the mainline kernel the clocks and resets are handled internally by the
media-blk-ctrl driver in power domains, while in the BSP driver they are
exposed to the consumers and have to be handled manually. Add support
for the BSP-specific clocks and resets to the driver.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
---
- Remove unused reset line