media: mc-entity: Add a new helper function to get a remote pad
The media_entity_remote_pad() helper function returns the first remote
pad it find connected to a given pad. Beside being possibly ill-named
(as it operates on a pad, not an entity) and non-deterministic (as it
stops at the first enabled link), the fact that it returns the first
match makes it unsuitable for drivers that need to guarantee that a
single link is enabled, for instance when an entity can process data
from one of multiple sources at a time.

For those use cases, add a new helper function,
media_entity_remote_pad_unique(), that operates on an entity and returns
a remote pad, with a guarantee that only one link is enabled. To ease
its use in drivers, also add an inline wrapper that locates source pads
specifically. A wrapper that locates sink pads can easily be added when
needed.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>