media: rkisp1: Add match data for i.MX8MP ISP
Add match data to the rkisp1 driver to match the i.MX8MP ISP.

Although the new version number isn't very precise, it ought to be fine
as the other version numbers aren't precise either, and we have separate
feature flags for important version-specific features. Despite this
version number being seemingly unimportant, it is added to distinguish
it from the ISP versions integrated in rockchip SoCs.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Rob Herring <robh@kernel.org>

---
Changes in v3:
- Remove todo for improving the version number
- Expand the commit message to address the version number