media: rkisp1: Add support for the TPG module on the i.MX8MP
Add support for the TPG module, specifically for the version of the ISP
on the i.MX8MP.

Note that there is no "disabled" option in the test pattern control, as
with the current design the TPG is disabled by disabling the link
between the TPG subdev and the ISP subdev.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
---
The validation for frame rates is not strict enough yet and some
selections (mainly edge cases) will result in no frames at one extreme
or missed irqs at the other extreme, but the average/common frame size
and frame rate this will work fine.