media: imx8-isi: Move hard-sleep from channel to pipe enable
The current implementation of the mxc_isi_channel_enable() function has
an hard 300 msec sleep, which comes from the BSP code but it is not
reported as required by the chip TRM.

Assuming the sleep has been inserted by the downstream developers based
on testing of the ISI, move it to the mxc_isi_pipe_enable() function
instead of removing it for the time being.

The reason for doing this is that when the ISI is used in m2m mode,
channels are enabled and disabled at each call of the v4l2 m2m framework
.device_run() callback, while the mxc_isi_pipe_enable() function is only
called at s_stream(1) time when the ISI is used in pixel capture mode
where such delay is acceptable.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>