media: rkisp1: Add support for 12MP
The ISP can support up to 12MP. Update the constants to reflect this.

The rk3399 TRM mentions a maximum size of 4416x3312 for the ISP (not
just the resizer), while the i.MX8MP only mentions 4096x3072. For now,
simply expand the maximum size to the minimum of these two, so that we
can support 12MP.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>