media: rkisp1: Support setting crop at runtime
Add support for setting the crop at runtime. Previously the crop could
only be set at configure time, and could not be changed at runtime. The
scaler component already worked properly, so this in effect enables
scaler crop.

Note that this is specifically for the i.MX8MP. As the rk3399 version
supports dual crop, it doesn't actually need to rely on the image
stabilizer for cropping. The rk3399 should still work with this, though,
but it needs to be taken into account when support for actual image
stabilization is added later.

media: rkisp1: Update resizer shadow registers automatically

The ISP (both the version in the rk3399 and the one in the i.MX8MP)
supports automatically updating the shadow registers of all subsequent
modules in the pipeline (resizer, memory interface, etc) when the last
pixel in a frame is finished. This alleviates the need for manually
timing when to update the shadow registers.

This feature was previously unused, causing an image size mismatch and
thus a pipeline stall when zooming in while doing digital zoom, where
the crop rectangle (in the image stabilizer, in the ISP) was changed
/before/ the scaling (in the resizer) was changed. That is, the image
size was shrunk by the ISP while the resizer still expected the old
bigger image size.

Fix this issue by changing the shadow register update mechanism to
auto-update when the ISP is done processing the frame.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>