media: rkisp1: Prevent race when setting digital zoom
When modifying the image stabilizer crop and resizer scaler for the
purposes of implementing digital zoom, this will change parameters on
both the ISP (for crop) and the resizer (for scaling), and changing
these two are not atomic. This could thus race, raising a wraparound
interrupt. Catch this, and mark the buffer as error.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>